requirements.txt
# TensorFlow for macOS
tensorflow-macos==2.12.0

# GPU acceleration for Apple Silicon (do NOT pin to exact version)
tensorflow-metal

# Scientific stack
numpy==1.23.5
pandas==1.5.3
matplotlib==3.7.1
scipy==1.10.1
requests==2.28.2
h5py==3.8.0
opencv-contrib-python==4.8.1.78
